Error handling for set compression

+++ comp-test.sh
@@ -1,22 +1,22 @@
 #!/bin/bash
 #Automated ZFS compressiontest
 
-BRANCH="master"
-git fetch
-git update-index -q --refresh
-CHANGED=$(git diff --name-only origin/$BRANCH)
-if [ ! -z "$CHANGED" ];
-then
-    echo "script requires update"
-    git reset --hard
-    git checkout $BRANCH
-    git pull
-    echo "script updated"
-    exit 1
-else
-    echo "script up-to-date"
-fi
-
+#BRANCH="master"
+#git fetch
+#git update-index -q --refresh
+#CHANGED=$(git diff --name-only origin/$BRANCH)
+#if [ ! -z "$CHANGED" ];
+#then
+#    echo "script requires update"
+#    git reset --hard
+#    git checkout $BRANCH
+#    git pull
+#    echo "script updated"
+#    exit 1
+#else
+#    echo "script up-to-date"
+#fi
+#
 
 now=$(date +%s)
 
@@ -217,22 +217,27 @@ then
         do
                 echo "running compression test for $comp"
                 ./zfs/cmd/zfs/zfs set compression=$comp testpool/fs1
-                echo “Compression results for $comp” >> "./$TESTRESULTS"
-                dd if=/mnt/ramdisk/$FILENAME of=/testpool/fs1/$FILENAME bs=4M  2>> "./$TESTRESULTS"
-                ./zfs/cmd/zfs/zfs get compressratio testpool/fs1 >> "./$TESTRESULTS"
-                echo "" >> "./$TESTRESULTS"
-                echo “Decompression results for $comp” >> "./$TESTRESULTS"
-                dd if=/testpool/fs1/$FILENAME of=/dev/null bs=4M  2>> "./$TESTRESULTS"
-                echo ""  >> "./$TESTRESULTS"
-                echo "verifying testhash"
-                cd /testpool/fs1/
-                chkresult=`echo "$chksum" | sha256sum --check`
-                sudo rm $FILENAME
-                cd -
-                echo "hashcheck result: $chkresult" >> "./$TESTRESULTS"
-                echo "" >> "./$TESTRESULTS"
-                echo "----" >> "./$TESTRESULTS"
-                echo "" >> "./$TESTRESULTS"
+		if [ $? -ne 0 ];
+		then
+			echo "Could not set compression to $comp! Skipping test."
+		else
+                	echo “Compression results for $comp” >> "./$TESTRESULTS"
+                	dd if=/mnt/ramdisk/$FILENAME of=/testpool/fs1/$FILENAME bs=4M  2>> "./$TESTRESULTS"
+                	./zfs/cmd/zfs/zfs get compressratio testpool/fs1 >> "./$TESTRESULTS"
+                	echo "" >> "./$TESTRESULTS"
+                	echo “Decompression results for $comp” >> "./$TESTRESULTS"
+                	dd if=/testpool/fs1/$FILENAME of=/dev/null bs=4M  2>> "./$TESTRESULTS"
+                	echo ""  >> "./$TESTRESULTS"
+                	echo "verifying testhash"
+                	cd /testpool/fs1/
+                	chkresult=`echo "$chksum" | sha256sum --check`
+                	sudo rm $FILENAME
+                	cd -
+                	echo "hashcheck result: $chkresult" >> "./$TESTRESULTS"
+                	echo "" >> "./$TESTRESULTS"
+                	echo "----" >> "./$TESTRESULTS"
+                	echo "" >> "./$TESTRESULTS"
+		fi
         done
 
         echo "compression test finished"
